+++ actionpack/test/controller/parameters/mutators_test.rb
@@ -27,24 +27,24 @@ class ParametersMutatorsTest < ActiveSupport::TestCase
   end
 
   test "delete returns the value when the key is present" do
-    assert_equal "32", @params[:person].delete(:age)
+    assert_equal "32", @params.param_at(:person).delete(:age)
   end
 
   test "delete removes the entry when the key present" do
-    @params[:person].delete(:age)
-    assert_not @params[:person].key?(:age)
+    @params.param_at(:person).delete(:age)
+    assert_not @params.param_at(:person).key?(:age)
   end
 
   test "delete returns nil when the key is not present" do
-    assert_nil @params[:person].delete(:first_name)
+    assert_nil @params.param_at(:person).delete(:first_name)
   end
 
   test "delete returns the value of the given block when the key is not present" do
-    assert_equal "David", @params[:person].delete(:first_name) { "David" }
+    assert_equal "David", @params.param_at(:person).delete(:first_name) { "David" }
   end
 
   test "delete yields the key to the given block when the key is not present" do
-    assert_equal "first_name: David", @params[:person].delete(:first_name) { |k| "#{k}: David" }
+    assert_equal "first_name: David", @params.param_at(:person).delete(:first_name) { |k| "#{k}: David" }
   end
 
   test "delete_if retains permitted status" do