+++ actionpack/test/controller/render_test.rb
@@ -115,16 +115,16 @@ class TestController < ActionController::Base
   end
 
   def dynamic_render
-    render params[:id] # => String, AC::Params
+    render params.param_at(:id) # => String, AC::Params
   end
 
   def dynamic_render_permit
-    render params[:id].permit(:file)
+    render params.param_at(:id).permit(:file)
   end
 
   def dynamic_render_with_file
     # This is extremely bad, but should be possible to do.
-    file = params[:id] # => String, AC::Params
+    file = params.param_at(:id) # => String, AC::Params
     render file: file
   end